uuid type (#12)
* uuid type
* NumberKind -> FloatKind
* ints and uints

+++ go/datas/database_test.go
@@ -83,7 +83,7 @@ func (suite *DatabaseSuite) TestCompletenessCheck() {
 
 	se := types.NewSet(suite.db).Edit()
 	for i := 0; i < 100; i++ {
-		se.Insert(suite.db.WriteValue(types.Number(100)))
+		se.Insert(suite.db.WriteValue(types.Float(100)))
 	}
 	s := se.Set()
 
@@ -91,7 +91,7 @@ func (suite *DatabaseSuite) TestCompletenessCheck() {
 	suite.NoError(err)
 
 	s = ds1.HeadValue().(types.Set)
-	s = s.Edit().Insert(types.NewRef(types.Number(1000))).Set() // danging ref
+	s = s.Edit().Insert(types.NewRef(types.Float(1000))).Set() // danging ref
 	suite.Panics(func() {
 		ds1, err = suite.db.CommitValue(ds1, s)
 	})
@@ -229,7 +229,7 @@ func (suite *DatabaseSuite) TestDatasetsMapType() {
 	suite.NotPanics(func() { assertMapOfStringToRefOfCommit(suite.db.Datasets(), datasets, suite.db) })
 
 	datasets = suite.db.Datasets()
-	_, err = suite.db.CommitValue(suite.db.GetDataset(dsID2), types.Number(42))
+	_, err = suite.db.CommitValue(suite.db.GetDataset(dsID2), types.Float(42))
 	suite.NoError(err)
 	suite.NotPanics(func() { assertMapOfStringToRefOfCommit(suite.db.Datasets(), datasets, suite.db) })
 
@@ -262,7 +262,7 @@ func (suite *DatabaseSuite) TestDatabaseCommitMerge() {
 	ds1, ds2 := suite.db.GetDataset(datasetID1), suite.db.GetDataset(datasetID2)
 
 	var err error
-	v := types.NewMap(suite.db, types.String("Hello"), types.Number(42))
+	v := types.NewMap(suite.db, types.String("Hello"), types.Float(42))
 	ds1, err = suite.db.CommitValue(ds1, v)
 	ds1First := ds1
 	suite.NoError(err)
@@ -273,11 +273,11 @@ func (suite *DatabaseSuite) TestDatabaseCommitMerge() {
 	suite.NoError(err)
 
 	// No common ancestor
-	_, err = suite.db.Commit(ds1, types.Number(47), newOpts(suite.db, ds2.HeadRef()))
+	_, err = suite.db.Commit(ds1, types.Float(47), newOpts(suite.db, ds2.HeadRef()))
 	suite.IsType(ErrMergeNeeded, err, "%s", err)
 
 	// Unmergeable
-	_, err = suite.db.Commit(ds1, types.Number(47), newOptsWithMerge(suite.db, merge.None, ds1First.HeadRef()))
+	_, err = suite.db.Commit(ds1, types.Float(47), newOptsWithMerge(suite.db, merge.None, ds1First.HeadRef()))
 	suite.IsType(&merge.ErrMergeConflict{}, err, "%s", err)
 
 	// Merge policies
@@ -289,10 +289,10 @@ func (suite *DatabaseSuite) TestDatabaseCommitMerge() {
 	suite.NoError(err)
 	suite.True(types.Bool(true).Equals(theirs.HeadValue().(types.Map).Get(types.String("Friends"))))
 
-	newV = v.Edit().Set(types.String("Friends"), types.Number(47)).Map()
+	newV = v.Edit().Set(types.String("Friends"), types.Float(47)).Map()
 	ours, err := suite.db.Commit(ds1First, newV, newOptsWithMerge(suite.db, merge.Ours, ds1First.HeadRef()))
 	suite.NoError(err)
-	suite.True(types.Number(47).Equals(ours.HeadValue().(types.Map).Get(types.String("Friends"))))
+	suite.True(types.Float(47).Equals(ours.HeadValue().(types.Map).Get(types.String("Friends"))))
 }
 
 func newOptsWithMerge(vrw types.ValueReadWriter, policy merge.ResolveFunc, parents ...types.Value) CommitOptions {
